app/requirements.txt
astroid==3.3.11
dill==0.4.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mysql-connector-python==9.4.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
tomlkit==0.13.3
pytest==8.4.1
